requirements-macos.txt
Pillow>=10.0
pyobjc-framework-Cocoa>=10.0
rumps>=0.4.0
pyinstaller>=6.0

